Added periodic stdout print statement to prevent buildbot from prematurely killing the tests

+++ fb-run.py
@@ -1,15 +1,25 @@
 from ConfigParser import ConfigParser
 from time import sleep
-import os, sys, signal, subprocess, shlex, getopt
+import os, sys, signal, subprocess, shlex, getopt, mozrunner
+
 
 def kill_firefox():
     for line in os.popen("ps xa"):
         fields = line.split()
         pid = fields[0]
         process = fields[4]
-        if process.find("firefox") > 0:
+        if process.find("firefox-bin") > 0 and process.find("<defunct>") == -1:
             os.kill(int(pid), signal.SIGHUP)
             break
+        
+def ping_firefox():
+    for line in os.popen("ps xa"):
+        fields = line.split()
+        pid = fields[0]
+        process = fields[4]
+        if process.find("firefox-bin") > 0 and process.find("<defunct>") == -1:
+            return 1
+    return 0
 
 # Initialization
 config = ConfigParser()
@@ -33,12 +43,13 @@ for o, a in opts:
     else:
         assert False, "unhandled option"
 
-
-
-serverpath = serverpath + "firebug" + firebug_version
+if serverpath[-1:] != "/":
+    serverpath = serverpath + "/firebug" + firebug_version
+else:
+    serverpath = serverpath + "firebug" + firebug_version
 
 # Grab the extensions
-os.system("wget " + serverpath + "/firebug.xpi " + serverpath + "/fbtest.xpi")
+os.system("wget -N " + serverpath + "/firebug.xpi " + serverpath + "/fbtest.xpi")
 
 # If firefox is running, kill it (needed for mozrunner)
 kill_firefox()
@@ -46,10 +57,19 @@ kill_firefox()
 sleep(2)
 
 subprocess.Popen(shlex.split("mozrunner -n -p " + profile + " --addons=\"./firebug.xpi\",\"./fbtest.xpi\""))
-os.system("firefox -runFBTests " + serverpath + "/tests/content/testlists/firebug" + firebug_version + ".html")# -P " + profile)
+sleep(5)
+subprocess.Popen("firefox -runFBTests " + serverpath + "/tests/content/testlists/firebug" + firebug_version + ".html", shell=True)
+
+count = 0
+while ping_firefox():
+    sleep(1)
+    count += 1
+    if count % 30 == 0:
+        print "I'm not dead yet!"
+
 
 # Cleanup
 os.system("rm ./firebug.xpi")
 os.system("rm ./fbtest.xpi")
-kill_firefox()
 sys.exit(0)
+
